APIS-4714 - Restructure packages in tests
