+++ test/uk/gov/hmrc/apidocumentation/controllers/DownloadControllerSpec.scala
@@ -22,8 +22,10 @@ import play.api.http.Status._
 import play.api.mvc._
 import uk.gov.hmrc.apidocumentation.ErrorHandler
 import uk.gov.hmrc.apidocumentation.controllers.utils._
+import uk.gov.hmrc.apidocumentation.mocks.services._
 import uk.gov.hmrc.apidocumentation.models.APIAccessType
 import uk.gov.hmrc.apidocumentation.services.DownloadService
+import uk.gov.hmrc.apidocumentation.mocks.config._
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
