+++ test/uk/gov/hmrc/apidocumentation/mocks/config/AppConfigMock.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.apidocumentation.controllers.utils
+package uk.gov.hmrc.apidocumentation.mocks.config
 
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
