+++ test/uk/gov/hmrc/apidocumentation/mocks/config/RamlPreviewEnabled.scala
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.apidocumentation.controllers.utils
-
+package uk.gov.hmrc.apidocumentation.mocks.config
 import org.mockito.Mockito._
 
 trait RamlPreviewEnabled {
