+++ test/uk/gov/hmrc/apidocumentation/mocks/connectors/ApiDefinitionConnectorMockingHelper.scala
@@ -14,59 +14,58 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.apidocumentation.utils
+package uk.gov.hmrc.apidocumentation.mocks.connectors
 
 import org.mockito.Matchers.{any, eq => eqTo}
 import org.mockito.Mockito.when
+import uk.gov.hmrc.apidocumentation.connectors.ApiDefinitionConnector
 import uk.gov.hmrc.apidocumentation.models.{APIDefinition, ExtendedAPIDefinition}
-import uk.gov.hmrc.apidocumentation.services.BaseApiDefinitionService
 import uk.gov.hmrc.http.HeaderCarrier
 
 import scala.concurrent.Future
 
-trait BaseApiDefinitionServiceMockingHelper {
-
-  def whenFetchAllDefinitions[T <: BaseApiDefinitionService](base: T)
+trait ApiDefinitionConnectorMockingHelper {
+  def whenFetchAllDefinitions[T <: ApiDefinitionConnector](base: T)
                              (apis: APIDefinition*)
                              (implicit hc: HeaderCarrier) = {
-    when(base.fetchAllDefinitions(any[None.type]())(eqTo(hc)))
+    when(base.fetchAllApiDefinitions(any[None.type]())(eqTo(hc)))
       .thenReturn(Future.successful(apis.toSeq))
   }
-  def whenFetchAllDefinitionsWithEmail[T <: BaseApiDefinitionService](base: T)
+  def whenFetchAllDefinitionsWithEmail[T <: ApiDefinitionConnector](base: T)
                                       (email: String)
                                       (apis: APIDefinition*)
                                       (implicit hc: HeaderCarrier) = {
-    when(base.fetchAllDefinitions(any[Some[String]]())(eqTo(hc)))
+    when(base.fetchAllApiDefinitions(any[Some[String]]())(eqTo(hc)))
       .thenReturn(Future.successful(apis.toSeq))
   }
-  def whenFetchExtendedDefinition[T <: BaseApiDefinitionService](base: T)
+  def whenFetchExtendedDefinition[T <: ApiDefinitionConnector](base: T)
                                  (serviceName: String)
                                  (api: ExtendedAPIDefinition)
                                  (implicit hc: HeaderCarrier) = {
-    when(base.fetchExtendedDefinition(eqTo(serviceName), eqTo(None))(eqTo(hc)))
+    when(base.fetchApiDefinition(eqTo(serviceName), eqTo(None))(eqTo(hc)))
       .thenReturn(Future.successful(Some(api)))
   }
-  def whenFetchExtendedDefinitionWithEmail[T <: BaseApiDefinitionService](base: T)
+  def whenFetchExtendedDefinitionWithEmail[T <: ApiDefinitionConnector](base: T)
                                           (serviceName: String, email: String)
                                           (api: ExtendedAPIDefinition)
                                           (implicit hc: HeaderCarrier) = {
-    when(base.fetchExtendedDefinition(eqTo(serviceName), any[Some[String]]())(eqTo(hc)))
+    when(base.fetchApiDefinition(eqTo(serviceName), any[Some[String]]())(eqTo(hc)))
       .thenReturn(Future.successful(Some(api)))
   }
 
-  def whenApiDefinitionFails[T <: BaseApiDefinitionService](base: T)
+  def whenApiDefinitionFails[T <: ApiDefinitionConnector](base: T)
                             (exception: Throwable)
                             (implicit hc: HeaderCarrier) = {
-    when(base.fetchExtendedDefinition(any[String],any[None.type]())(eqTo(hc)))
+    when(base.fetchApiDefinition(any[String],any[None.type]())(eqTo(hc)))
       .thenReturn(Future.failed(exception))
-    when(base.fetchAllDefinitions(any())(eqTo(hc)))
+    when(base.fetchAllApiDefinitions(any())(eqTo(hc)))
       .thenReturn(Future.failed(exception))
   }
 
-  def whenNoApiDefinitions[T <: BaseApiDefinitionService](base: T) = {
-    when(base.fetchExtendedDefinition(any[String],any())(any[HeaderCarrier]))
+  def whenNoApiDefinitions[T <: ApiDefinitionConnector](base: T)(implicit hc: HeaderCarrier) = {
+    when(base.fetchApiDefinition(any[String],any())(eqTo(hc)))
       .thenReturn(Future.successful(None))
-    when(base.fetchAllDefinitions(any())(any[HeaderCarrier]))
+    when(base.fetchAllApiDefinitions(any())(eqTo(hc)))
       .thenReturn(Future.successful(Seq.empty))
   }
 }
